% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.addMetadataForAggregates}
\alias{.addMetadataForAggregates}
\title{Adding matrix metadata.}
\usage{
.addMetadataForAggregates(
  ArchRProj,
  o,
  peakSet,
  knnObj,
  useMatrix,
  gS,
  log2Norm,
  scaleTo
)
}
\description{
The function adds closest cells to seeds and normalized matrix as metadata to the analysis result.
}
\keyword{internal}
