% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.checkInputParameters}
\alias{.checkInputParameters}
\title{Input parameter check.}
\usage{
.checkInputParameters(parameterList)
}
\description{
The function takes the list of possible parameters and checks the input type.
}
\keyword{internal}
