% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.checkNumAggregatesAndCellsPerAggregate}
\alias{.checkNumAggregatesAndCellsPerAggregate}
\title{Cell number parameter check.}
\usage{
.checkNumAggregatesAndCellsPerAggregate(
  AggregationMethod,
  numAggregates,
  numCellsPerAggregate,
  cells_number
)
}
\description{
The function checks whether there are enough cells and adapts parameters if needed.
}
\keyword{internal}
