% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.createGroupMatrix}
\alias{.createGroupMatrix}
\title{Get and normalize Group Matrix.}
\usage{
.createGroupMatrix(
  ArchRProj,
  peakSet,
  knnObj,
  useMatrix,
  gS,
  log2Norm,
  chr,
  scaleTo
)
}
\description{
The function extracts the matrix for the given features and scales/normalizes it.
}
\keyword{internal}
