% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.createLoopsSameChromosome}
\alias{.createLoopsSameChromosome}
\title{Create Loops between Peaks.}
\usage{
.createLoopsSameChromosome(coA)
}
\description{
The function creates loops between pairs of peaks from peak indices.
}
\keyword{internal}
