% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.createPairwiseThingsToTest}
\alias{.createPairwiseThingsToTest}
\title{Pairwise list of features.}
\usage{
.createPairwiseThingsToTest(peakSet, maxDist)
}
\description{
The function creates the pairwise list of features for the co-accessibility analysis.
}
\keyword{internal}
