% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.filterKNN}
\alias{.filterKNN}
\title{Filter KNN object.}
\usage{
.filterKNN(
  knnObj,
  AggregationMethod,
  overlapCutoff,
  numCellsPerAggregate,
  numAggregates
)
}
\description{
The function calculates the overlap between KNNs of each seed cell and keeps those with more than overlapCutoff unique cells in KNN.
It does not apply to single cell level mode of aggregation, because we want to keep all cells.
}
\keyword{internal}
