% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.getColSumsCustomMatrix}
\alias{.getColSumsCustomMatrix}
\title{Get ColSums for custom matrix.}
\usage{
.getColSumsCustomMatrix(customMatrix)
}
\description{
The function substitutes .getColSums from ArchR.
}
\keyword{internal}
