% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.getFilteredReducedDimensions}
\alias{.getFilteredReducedDimensions}
\title{Reduced Dimensions getter.}
\usage{
.getFilteredReducedDimensions(
  ArchRProj,
  reducedDims,
  corCutOff,
  dimsToUse,
  cellsToUse
)
}
\description{
The function gets the reduced dimensions and, if specified, filters the cells.
}
\keyword{internal}
