% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.getSet}
\alias{.getSet}
\title{Feature Set getter}
\usage{
.getSet(ArchRProj, useMatrix, binaryMatrix)
}
\description{
The function returns feature set based on the specified matrix.
}
\keyword{internal}
