% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.selectCellSeedsForAggregation}
\alias{.selectCellSeedsForAggregation}
\title{Cell seeds selection.}
\usage{
.selectCellSeedsForAggregation(
  ArchRProj,
  reducedDimensions,
  AggregationMethod,
  numPermutations,
  numCellsPerAggregate,
  numAggregates,
  cellsToUse
)
}
\description{
The function creates the list of cells on which the co-accessibility analysis is performed.
Depends on the AggregationMethod parameter.
}
\keyword{internal}
