% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coAccessibilityHelper.R
\name{.selectClosestCellsOfCellSeeds}
\alias{.selectClosestCellsOfCellSeeds}
\title{Closest cells to seeds selection.}
\usage{
.selectClosestCellsOfCellSeeds(
  ArchRProj,
  reducedDimensions,
  idx,
  AggregationMethod,
  numAggregates,
  numCellsPerAggregate
)
}
\description{
The function performs KNN to find the closest cells to seeds.
}
\keyword{internal}
