% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellAggregateMatrix.R
\name{getCellAggregateMatrix}
\alias{getCellAggregateMatrix}
\title{getBackgroundCoAccessibility}
\usage{
getCellAggregateMatrix(
  ArchRProj = NULL,
  reducedDims = "IterativeLSI",
  k = 100,
  knnIteration = 500,
  corCutOff = 0.75,
  dimsToUse = 1:30,
  overlapCutoff = 0.8,
  scaleTo = 10^4
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{k}{The number of k-nearest neighbors to use for creating single-cell groups for correlation analyses. Number of cells per aggregate.}

\item{knnIteration}{The number of k-nearest neighbor groupings to test for passing the supplied \code{overlapCutoff}. Number of aggregates.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to
sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{overlapCutoff}{The maximum allowable overlap between the current group and all previous groups to permit the current group be
added to the group list during k-nearest neighbor calculations.}

\item{scaleTo}{The total insertion counts from the designated group of single cells is summed across all relevant peak regions from
the \code{peakSet} of the \code{ArchRProject} and normalized to the total depth provided by \code{scaleTo}.}
}
\description{
This function returns background co-accessibility scores for peaks of a given ArchRProject. The background is simulated by shuffling of features and cells.
}
\examples{
Coming soon.
}
\keyword{aggregates}
\keyword{archr}
\keyword{archrwire}
\keyword{matrix}
\keyword{rwire}
