% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoAccessibility.R
\name{getCoAccessibility}
\alias{getCoAccessibility}
\title{Get Peak Co-Accessibility to an ArchRProject}
\usage{
getCoAccessibility(
  ArchRProj = NULL,
  reducedDims = "IterativeLSI",
  dimsToUse = 1:30,
  scaleDims = NULL,
  corCutOff = 0.75,
  cellsToUse = NULL,
  AggregationMethod = "ArchR_default",
  numCellsPerAggregate = 100,
  numAggregates = 500,
  useMatrix = "PeakMatrix",
  overlapCutoff = 0.8,
  maxDist = 1e+05,
  scaleTo = 10^4,
  log2Norm = TRUE,
  seed = 1,
  threads = getArchRThreads(),
  numPermutations = 1000,
  verbose = TRUE,
  logFile = createLogFile("getCoAccessibility")
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{dimsToUse}{A vector containing the dimensions from the \code{reducedDims} object to use in clustering.}

\item{scaleDims}{A boolean value that indicates whether to z-score the reduced dimensions for each cell. This is useful for minimizing
the contribution of strong biases (dominating early PCs) and lowly abundant populations. However, this may lead to stronger sample-specific
biases since it is over-weighting latent PCs. If set to \code{NULL} this will scale the dimensions based on the value of \code{scaleDims} when the
\code{reducedDims} were originally created during dimensionality reduction. This idea was introduced by Timothy Stuart.}

\item{corCutOff}{A numeric cutoff for the correlation of each dimension to the sequencing depth. If the dimension has a correlation to
sequencing depth that is greater than the \code{corCutOff}, it will be excluded from analysis.}

\item{cellsToUse}{A character vector of cellNames to compute coAccessibility on if desired to run on a subset of the total cells.}

\item{AggregationMethod}{Choose cell aggregation method from "ArchR_default" (each cell in multiple aggregates), "unique" (each cell in only one aggregate), or "single_cell_resolution" (no cell aggregation, single-cell level).}

\item{numCellsPerAggregate}{The number of k-nearest neighbors to use for creating single-cell groups for correlation analyses.}

\item{numAggregates}{The number of k-nearest neighbor groupings to test for passing the supplied \code{overlapCutoff}.}

\item{useMatrix}{The name of the data matrix to use for calculation of CoAccessibility. Options include "PeakMatrix" or non-binarized "TileMatrix" (binarize = FALSE in addTileMatrix).}

\item{overlapCutoff}{The maximum allowable overlap between the current group and all previous groups to permit the current group be
added to the group list during k-nearest neighbor calculations.}

\item{maxDist}{The maximum allowable distance in basepairs between two peaks to consider for co-accessibility.}

\item{scaleTo}{The total insertion counts from the designated group of single cells is summed across all relevant peak regions from
the \code{featureSet} of the \code{ArchRProject} and normalized to the total depth provided by \code{scaleTo}.}

\item{log2Norm}{A boolean value indicating whether to log2 transform the single-cell groups prior to computing co-accessibility correlations.}

\item{seed}{A number to be used as the seed for random number generation required in knn determination. It is recommended to keep track
of the seed used so that you can reproduce results downstream.}

\item{threads}{The number of threads to be used for parallel computing.}

\item{numPermutations}{The number of permutations for computing a distant group of random cells.
Increasing this number will not necessarily lead to a more distant group of cells.}

\item{verbose}{A boolean value that determines whether standard output should be printed.}

\item{logFile}{The path to a file to be used for logging ArchR output.}
}
\description{
This function is an extended version of ArchR package that will add co-accessibility scores between peaks in a given ArchRProject.
There are two new modes of choosing cell aggregates: unique and single_cell_resolution. If you are looking for co-accessibility in homogeneous population,
where the difference in accessibility is explained by stochastic nature rather than external influence, consider "single_cell_resolution" mode. There is no aggregation in this mode, hence single cell resolution.
Either all cells or only the cells specified in cellsToUse are used.
"Unique" mode works similar to the default of ArchR but only allows each cell to be in one aggregate.
}
