% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinkActivityMatrix.R
\name{getLinkActivityMatrix}
\alias{getLinkActivityMatrix}
\title{Get Link Activity Matix}
\usage{
getLinkActivityMatrix(
  ArchRProj,
  CoAccessibility,
  useMatrix = "PeakMatrix",
  peakSubset = NULL,
  regionSubset = NULL,
  binarize = TRUE,
  correlationCoefficient = FALSE,
  logFile = createLogFile("getLinkActivityMatrix"),
  verbose = TRUE
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{CoAccessibility}{The co-accessibility data produced by RWireX method getCoAccessibility.}

\item{useMatrix}{Same matrix type that was used for the accessibility.}

\item{peakSubset}{If specified, then the matrix is subsetted by the links with the first peak in the peakSubset.}

\item{regionSubset}{If specified, then the matrix is subsetted by the links where the first peak is in this range.}

\item{binarize}{if matrix needs to be binarized. Have to be the same as the matrix used for CoAccessibility calculation.}

\item{correlationCoefficient}{multiply links by correlation.}

\item{logFile}{The path to a file to be used for logging RWireX output.}

\item{verbose}{A boolean value that determines whether standard output should be printed.}
}
\description{
Get Link Activity Matix
}
