% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellQuality.R
\name{plotCellQuality}
\alias{plotCellQuality}
\title{plotCellQuality}
\usage{
plotCellQuality(
  ArchRProj,
  sample,
  groups = "Sample",
  cutoff_frags = NULL,
  cutoff_tss = NULL,
  set_xlim = NULL,
  set_ylim = NULL
)
}
\arguments{
\item{ArchRProj}{ArchRProject.}

\item{sample}{Sample name.}

\item{groups}{CellColData column to group samples by.}

\item{cutoff_frags}{Optional additional cutoff for TSS Enrichment to visualize.}

\item{set_xlim}{Optional limits for x-axis (Log10 Unique Fragments).}

\item{set_ylim}{Optional limits for y-axis (TSS Enrichment).}
}
\value{
GGplot object of scatter plot (TSS Enrichment over log10 Unique Fragemnts).
}
\description{
This function returns a ggplot object with a scatter plot of TSS Enrichment over log10 Unique Fragments per cell. Default filtering criteria of TSS Enrichment of 4 and minimal number of unique fragments of 1000 are visualized.
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{archrwire}
\keyword{cell}
\keyword{qc}
\keyword{quality}
\keyword{rwire}
