% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoAccessibilityMap.R
\name{plotCoAccessibilityMap}
\alias{plotCoAccessibilityMap}
\title{plotCoAccessibilityMap}
\usage{
plotCoAccessibilityMap(
  ArchRProj = NULL,
  reducedDims = names(ArchRProj@reducedDims)[1],
  region = NULL,
  resolution = 1000,
  ArchRGenome = getArchRGenome(),
  cor_max = NULL,
  aggregation_k = 100,
  aggregation_knnIteration = 500,
  aggregation_overlapCutoff = 0.8
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{reducedDims}{The name of the \code{reducedDims} object (i.e. "IterativeLSI") to retrieve from the designated \code{ArchRProject}.}

\item{region}{GRanges object of length 1 with region of interest for which co-accessibility should be calculated.}

\item{resolution}{Tile size for which co-accessibility should be calculated. Defaults to 1 kb.}

\item{ArchRGenome}{Used genome and gene annotation.}

\item{aggregation_k}{The number of k-nearest neighbors to use for creating single-cell groups for correlation analyses (number of cells per aggregate).}

\item{aggregation_knnIteration}{The number of k-nearest neighbor groupings to test for passing the supplied overlapCutoff (number of aimed cell aggregates).}

\item{aggregation_overlapCutoff}{The maximum allowable overlap between the current group and all previous groups to permit the current group be added to the group list during k-nearest neighbor calculations.}
}
\value{
Co-accessibility map, GRanges object of co-accessibility scores, and tile set of region.
}
\description{
This function returns co-accessibility map of specified genomic region.
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{archrwire}
\keyword{co-accessibility}
\keyword{coAccessibilityMap}
\keyword{map}
\keyword{rwire}
