% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureViolin.R
\name{plotFeatureViolin}
\alias{plotFeatureViolin}
\title{Create a violin plot of cell values per group, e.g. samples.}
\usage{
plotFeatureViolin(
  ArchRProj,
  feature,
  useMatrix = "PeakMatrix",
  binarize = FALSE,
  matrix = NULL,
  groupBy = "Sample",
  normBy = "nFrags",
  orderedSamples = NULL,
  useGroups = NULL,
  pseudolog = FALSE,
  pal = NULL
)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{feature}{Peak to plot in the format chr:idx, for example "chr1:15" if
you use PeakMatrix, gene name if you use GeneScoreMatrix, or one of the data columns names if you use cellColData.}

\item{useMatrix}{Type of Matrix to plot data from}

\item{binarize}{If you plot values from a binary matrix. MUST be specified if the matrix is binarized.}

\item{matrix}{You can provide matrix directly for a faster run.}

\item{groupBy}{Which groups should be displayed in the plot.}

\item{normBy}{Specify if you want the values in the matrix normalized.}

\item{orderedSamples}{You can specify the order of the group in the plot here in a string vector.}

\item{useGroups}{Filter the groups.}

\item{pseudolog}{If you would like the values to be plotted in the pseudolog.}

\item{pal}{Color pallete}
}
\description{
This function returns a plot.
}
