% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUmapQC.R
\name{plotUmapQC}
\alias{plotUmapQC}
\title{plotUmapQC}
\usage{
plotUmapQC(
  ArchRProj,
  embedding = names(ArchRProj@embeddings)[1],
  coloringLayer = "Sample",
  title = "",
  coloringPalette = NULL,
  doubletScore = FALSE
)
}
\arguments{
\item{ArchRProj}{ArchRProject with priorly calculated reduced dimensions and single-cell embedding.}

\item{embedding}{Optional name of calculated single-cell embedding (e. g. calculated with addUMAP function). Defaults to first embedding in ArchRProj@embeddings.}

\item{coloringLayer}{Optional name of cellColData column to use as comparative coloring layer of single cells. Defaults to 'Sample' column.}

\item{title}{Optional title of quality control UMAPs.}

\item{coloringPalette}{Optional coloring palette for reference coloring of single cells.}

\item{doubletScore}{Optional addition of doublet score as QC metric.}
}
\value{
List of UMAPs. First with comparative coloring layer for quality assessment of single cells. Other UMAPs colored by total number of fragments, blacklist ratio, TSS enrichment, promoter ratio, nucleosome ratio, and if set doublet score.
}
\description{
This function returns a list of UMAPs with cells colored by various quality control meassures (total number of fragments, blacklist ratio, TSS enrichment, promoter ratio, nucleosome ratio).
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{archrwire}
\keyword{qc}
\keyword{quality}
\keyword{rwire}
\keyword{umap}
