% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUmaps.R
\name{plotUmaps}
\alias{plotUmaps}
\title{plotUmaps}
\usage{
plotUmaps(
  ArchRProj,
  embeddings = names(ArchRProj@embeddings),
  coloringLayer = "Sample",
  coloringPalette = NULL,
  size = 0.3
)
}
\arguments{
\item{ArchRProj}{ArchRProject with priorly calculated reduced dimensions and single-cell embedding.}

\item{embeddings}{Optional character vector of calculated single-cell embeddings (e. g. calculated with addUMAP function). Defaults to all ArchRProj@embeddings.}

\item{coloringLayer}{Optional name of cellColData column to use as coloring layer of single cells. Defaults to 'Sample' column.}

\item{coloringPalette}{Optional coloring palette for split levels of single cells.}

\item{size}{Point size of cells in UMAP embedding.}
}
\value{
List of UMAPs with different embeddings of single cells.
}
\description{
This function returns a list of UMAPs of different single-cell embeddings.
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{archrwire}
\keyword{embeddings}
\keyword{rwire}
\keyword{umap}
