% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUmapSplit.R
\name{plotUmapSplit}
\alias{plotUmapSplit}
\title{plotUmapSplit}
\usage{
plotUmapSplit(
  ArchRProj,
  embedding = names(ArchRProj@embeddings)[1],
  splitLayer = "Sample",
  selectLevels = unique(ArchRProj$Sample),
  coloringPalette = NULL
)
}
\arguments{
\item{ArchRProj}{ArchRProject with priorly calculated reduced dimensions and single-cell embedding.}

\item{embedding}{Optional name of calculated single-cell embedding (e. g. calculated with addUMAP function). Defaults to first embedding in ArchRProj@embeddings.}

\item{splitLayer}{Optional name of cellColData column to use as split layer of single cells. Defaults to 'Sample' column.}

\item{selectLevels}{Optional selection of certain levels in split layer to visualize. Defaults to all levels in cellColData 'Sample' column.}

\item{coloringPalette}{Optional coloring palette for split levels of single cells.}
}
\value{
List of UMAPs with cells of single split levels.
}
\description{
This function returns a list of UMAPs with cells of single split levels.
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{archrwire}
\keyword{rwire}
\keyword{split}
\keyword{umap}
