% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinkActivityMatrix.R
\name{getLinkActivityMatrix}
\alias{getLinkActivityMatrix}
\title{Get Link Activity Matix}
\usage{
getLinkActivityMatrix(ArchRProj, coAccessibility)
}
\arguments{
\item{ArchRProj}{An \code{ArchRProject} object.}

\item{coAccessibility}{The co-accessibility data produced by RWireX method getCoAccessibility.}
}
\description{
Get Link Activity Matix
}
