% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoAccessibilityMap.R
\name{plotCoAccessibilityMap}
\alias{plotCoAccessibilityMap}
\title{plotCoAccessibilityMap}
\usage{
plotCoAccessibilityMap(
  CoAccessibility,
  region,
  genome = ArchR::getArchRGenome(),
  features = NULL,
  main = NULL,
  onlyPos = FALSE,
  scaleLim = NULL,
  rescale = FALSE,
  annotation = NULL,
  annotation_size = 0.5
)
}
\arguments{
\item{CoAccessibility}{A GRanges object with co-accessible links from getCoAccessibility function.}

\item{region}{GRanges object of length 1 with region of interest for which co-accessibility should be visualized.}

\item{genome}{Genome and gene annotation to use.}

\item{main}{Title of co-accessibility map.}

\item{onlyPos}{Only visualize positive co-accessibility scores.}

\item{scaleLim}{Set limits of co-accessibility scale.}

\item{rescale}{If TRUE, genomic coordinate is rescaled to visualize peaks continously.}

\item{useMatrix}{The name of the data matrix used for calculation of CoAccessibility. Options include "PeakMatrix" or non-binarized "TileMatrix" (binarize = FALSE in addTileMatrix).}
}
\value{
Co-accessibility map
}
\description{
This function returns co-accessibility map of a specified genomic region.
}
\examples{
Coming soon.
}
\keyword{archr}
\keyword{co-accessibility}
\keyword{coAccessibilityMap}
\keyword{map}
\keyword{rwire}
